requirements.txt
wget
asyncio
pyrogram==1.0.1
tgcrypto
google-auth-oauthlib>=1.0.0
psycopg2-binary
sqlalchemy==1.3.23
httplib2
pySmartDL
google-api-python-client>=2.50.0
google-auth-httplib2
google-auth>=2.0.0
youtube_dl
ffmpeg
tenacity
html5lib
lk21
beautifulsoup4
lxml
